+++ src/screens/Sidebar/Blocking.js
@@ -6,9 +6,7 @@ import PropTypes from 'prop-types';
 import * as colors from 'kitsu/constants/colors';
 import menu from 'kitsu/assets/img/tabbar_icons/menu.png';
 
-import SidebarHeader from './common/SidebarHeader';
-import SidebarTitle from './common/SidebarTitle';
-import { ItemSeparator } from './common/SidebarListItem';
+import { SidebarHeader, SidebarTitle, ItemSeparator } from './common/';
 
 class Blocking extends React.Component {
   static navigationOptions = ({ navigation }) => ({
